Split MiniPlaceholders in modules
- Added Paper module
Requires Paper 1.18.2+
