+++ api/src/main/java/me/dreamerzero/miniplaceholders/api/placeholder/ExpansionPlaceholder.java
@@ -1,0 +1,19 @@
+package me.dreamerzero.miniplaceholders.api.placeholder;
+
+import org.jetbrains.annotations.NotNull;
+
+abstract class ExpansionPlaceholder<F> {
+    private String name;
+    private F function;
+    protected ExpansionPlaceholder(String name, F function){
+        this.name = name;
+        this.function = function;
+    }
+    public @NotNull String name() {
+        return this.name;
+    }
+
+    public @NotNull F get() {
+        return this.function;
+    }
+}
