+++ paper/src/main/java/me/dreamerzero/miniplaceholders/paper/PaperPlugin.java
@@ -1,0 +1,10 @@
+package me.dreamerzero.miniplaceholders.paper;
+
+import org.bukkit.plugin.java.JavaPlugin;
+
+public class PaperPlugin extends JavaPlugin {
+    @Override
+    public void onEnable(){
+        this.getSLF4JLogger().info("Starting MiniPlaceholders Paper");
+    }
+}
